Adding Hitbox and Game classes, adjusting combolyzer to use new class model, etc.
I added a new Hitbox class, which I had not originally planned. However,
I feel now that it is probably better to separate the abstractions of
hitbox and action. Hitboxes are complicated, and actions may have many
hitboxes. That said, I'm not totally confident in how I currently have
`Action` setup (with just an array of hitboxes and a corresponding array
of active frames). I may end up finding a way to parameterize move
scripts, although that's probably overkill. For now, this gives proper
results despite all the changes, so I'm fine with committing it.

Fighter received one new property: `overrides`. I intend to add an
`overrides` property to actions and possible hitboxes as well. The idea
is that for valules in each formula, there may be a corresponding
property on the `overrides` object of a fighter or action. This property
points to a function which accepts the default value (the normal value
of that property for most characters) and some other information, and
returns a modified version of that value.

Lucario's aura is currently implemented as the first and only override
at the moment. Because aura affects the base damage of Lucario's moves,
Lucario gets the override `Lucario.overrides.baseDamage`, which points
to a function that hands the appropriate values to the Aura function
that Ruben implemented. This is currently only used in VSKB. Other
overrides may exist in the future. For example, Bowser may get a
knockback and hitstun override corresponding to the Tough Guy effect.
Cloud may get a weight and fallspeed override for Limit Break. Etc.

`Game` is a pretty barebones class at the moment. Just containing an
array of fighters in the battle, which is used in the aura calculation
to find who has the most stocks to compare to Lucario's stock count.